Launch dnsmasq on startup
Proxy historically has not had its own dnsmasq as we need to use dnsmasq
over TCP (not UDP). This is because of a kernel network routing bug
w.r.t udp connections where UDP packets fail to be delivered when a
container is restarted a large number of times.

We make sure we to use TCP for dns lookups by using the 'use-vc' option in
RES_OPTIONS. This is however only for libraries using glibc. Nginx does
not use glibc for resolving names. Therefore we used a dnsmasq running
in another container which we accessed through the gateway IP. This does
not work in Fedora 23, because Fedora 23 has strange rules where
communicating between 2 docker containers over the gateway over UDP for
ports < 1024 fail.

So we have mapped dnsmasq on another port > 1024, and are using that
port.