fix upload problems
With the migration to AlpineLinux, /var/lib/nginx seems to be owned by
root, and then nginx receives permission errors when trying to upload
files, as it stores them temporally inside /var/lib/nginx, and the nginx
user does not have permission there.

Changing the owner of /var/lib/nginx to nginx fixes the problem.