Adds feature toggling for serf
We need to modify start.sh script to detect if we want to use serf as
the service discovery. In such case, we run proxy as normal.

Otherwhise, we need to modify the gateway hostname from the nginx.conf
file to localhost. Due to this change we need to run locally dnsmasq and
run nginx with eyeos-run-server.